Refactor to reduce duplicity
