+++ src/Tooltip.jsx
@@ -1,44 +1,42 @@
 /**
  * @class Tooltip
- * @description A lightweight and responsive tooltip.
+ * @description The container of a lightweight and responsive tooltip.
  */
 import React from 'react';
 import PropTypes from 'prop-types';
 
-import Portal from 'react-minimalist-portal';
-import positions from './position';
+import TooltipBubble from './TooltipBubble';
 
-// default colors
-const defaultColor = '#fff';
-const defaultBg = '#333';
-
-const stopProp = e => e.stopPropagation();
-
-class Tooltip extends React.PureComponent {
+class Tooltip extends React.Component {
   static propTypes = {
+    // eslint-disable-next-line react/no-unused-prop-types
     direction: PropTypes.string,
-    className: PropTypes.string,
+    tooltipClassName: PropTypes.string,
     content: PropTypes.node.isRequired,
     background: PropTypes.string,
     color: PropTypes.string,
     padding: PropTypes.string,
+    eventOff: PropTypes.string,
+    eventOn: PropTypes.string,
     eventToggle: PropTypes.string,
     useHover: PropTypes.bool,
     useDefaultStyles: PropTypes.bool,
+    isOpen: PropTypes.bool,
+    hoverDelay: PropTypes.number,
     tipContentHover: PropTypes.bool,
     arrow: PropTypes.bool,
     arrowSize: PropTypes.number,
     distance: PropTypes.number,
-    target: PropTypes.object,
   };
   static defaultProps = {
     direction: 'up',
-    className: '',
+    tooltipClassName: '',
     background: '',
     color: '',
     padding: '10px',
     useHover: true,
     useDefaultStyles: false,
+    hoverDelay: 200,
     tipContentHover: false,
     arrow: true,
     arrowSize: 10,
@@ -47,82 +45,178 @@ class Tooltip extends React.PureComponent {
 
   constructor() {
     super();
+
+    this.state = {
+      showTip: false,
+      hasHover: false,
+    };
+
+    this.target = React.createRef();
     this.tip = React.createRef();
+
+    this.showTip = this.showTip.bind(this);
+    this.hideTip = this.hideTip.bind(this);
+    this.checkHover = this.checkHover.bind(this);
+    this.toggleTip = this.toggleTip.bind(this);
+    this.startHover = this.startHover.bind(this);
+    this.endHover = this.endHover.bind(this);
   }
 
   componentDidMount() {
-    // This is necessary because positions relies on the tooltip existing to determine the optimal location
-    this.forceUpdate();
+    // if the isOpen prop is passed on first render we need to immediately trigger a second render,
+    // because the tip ref is needed to calculate the position
+    if (this.props.isOpen) {
+      // eslint-disable-next-line react/no-did-mount-set-state
+      this.setState({ isOpen: true });
+    }
   }
 
+  target = null;
   tip = null;
+
+  toggleTip() {
+    this.setState({ showTip: !this.state.showTip });
+  }
+
+  showTip() {
+    this.setState({ showTip: true });
+  }
+
+  hideTip() {
+    this.setState({ hasHover: false });
+    this.setState({ showTip: false });
+  }
+
+  startHover() {
+    this.setState({ hasHover: true });
+
+    setTimeout(this.checkHover, this.props.hoverDelay);
+  }
+
+  endHover() {
+    this.setState({ hasHover: false });
+
+    setTimeout(this.checkHover, this.props.hoverDelay);
+  }
+
+  checkHover() {
+    this.setState({ showTip: this.state.hasHover });
+  }
 
   render() {
     const {
       direction,
-      className,
+      tooltipClassName,
       padding,
+      children,
       content,
+      eventOn,
+      eventOff,
       eventToggle,
       useHover,
       background,
       color,
       useDefaultStyles,
+      isOpen,
       tipContentHover,
       arrow,
       arrowSize,
       distance,
-      target,
+      ...others
     } = this.props;
+    delete others.hoverDelay;
 
-    const currentPositions = positions(direction, this.tip.current, target, {
-      background: useDefaultStyles ? defaultBg : background,
-      arrow,
-      arrowSize,
-      distance,
-    });
+    const showTip = (typeof isOpen === 'undefined') ? this.state.showTip : isOpen;
 
-    const tipStyles = {
-      ...currentPositions.tip,
-      background: useDefaultStyles ? defaultBg : background,
-      color: useDefaultStyles ? defaultColor : color,
-      padding,
-      boxSizing: 'border-box',
-      zIndex: 1000,
-      position: 'absolute',
-      display: 'inline-block',
-    };
+    const props = {};
 
-    const arrowStyles = {
-      ...currentPositions.arrow,
-      position: 'absolute',
-      width: '0px',
-      height: '0px',
-      zIndex: 1001,
-    };
-
-    const portalProps = {
-      // keep clicks on the tip from closing click controlled tips
-      onClick: stopProp,
-    };
-
-    if (!eventToggle && useHover && tipContentHover) {
-      portalProps.onMouseOver = this.startHover;
-      portalProps.onMouseOut = this.endHover;
-      portalProps.onTouchStart = stopProp;
+    // event handling
+    if (eventOff) {
+      props[eventOff] = this.hideTip;
     }
 
+    if (eventOn) {
+      props[eventOn] = this.showTip;
+    }
+
+    if (eventToggle) {
+      props[eventToggle] = this.toggleTip;
+
+      // only use hover if they don't have a toggle event
+    } else if (useHover) {
+      props.onMouseOver = this.startHover;
+      props.onMouseOut = tipContentHover ? this.endHover : this.hideTip;
+      props.onTouchStart = this.toggleTip;
+    }
+
+    // should ensure a single child (throws an error instead)
+    React.Children.only(children);
+    // map other properties and most importantly, reference to the inner DOM component
+    const updatedChildren = React.Children.map(children, (child) => {
+      const additionalProps = {
+        ...props,
+        ...others,
+      };
+      if (typeof child.type === 'function') {
+        // if the Tooltip is attaching to another React Component
+        // the inner React Component MUST handle the passing of the ref on its own (as well as spread other props (most importantly the events)
+        return React.cloneElement(child, {
+          innerRef: this.target,
+          ...additionalProps,
+        });
+      } else {
+        // or an HTML node
+        return React.cloneElement(child, {
+          ref: this.target,
+          ...additionalProps,
+        });
+      }
+    });
     return (
-      <Portal>
-        <div {...portalProps} className={className}>
-          <span className="react-tooltip-lite" style={tipStyles} ref={this.tip}>
-            {content}
-          </span>
-          <span className={`react-tooltip-lite-arrow react-tooltip-lite-${currentPositions.realDirection}-arrow`} style={arrowStyles} />
-        </div>
-      </Portal>
+      <React.Fragment>
+        {updatedChildren}
+        {showTip && (
+          <TooltipBubble
+            direction={direction}
+            className={tooltipClassName}
+            content={content}
+            background={background}
+            color={color}
+            padding={padding}
+            eventToggle={eventToggle}
+            useHover={useHover}
+            useDefaultStyles={useDefaultStyles}
+            tipContentHover={tipContentHover}
+            arrow={arrow}
+            arrowSize={arrowSize}
+            distance={distance}
+            target={this.target.current}
+          />
+        )}
+      </React.Fragment>
     );
   }
 }
 
-export default Tooltip;
+const TooltipElement = ({ content, children, wrapperProps, ...others }) => (
+  <Tooltip
+    content={content}
+    {...others}
+  >
+    <dfn {...wrapperProps}>
+      {children}
+    </dfn>
+  </Tooltip>
+);
+
+TooltipElement.propTypes = {
+  content: PropTypes.node.isRequired,
+  children: PropTypes.node.isRequired,
+  wrapperProps: PropTypes.object,
+};
+
+TooltipElement.defaultProps = {
+  wrapperProps: {},
+};
+
+export { Tooltip, TooltipElement };
